Fix logout redirects (#2704)
